Refactor: changed cards_temp to cards
