+++ src/app/docs/animations/hovers/jiggle/page.mdx
@@ -1,6 +1,6 @@
 import Link from 'next/link'
 import { RequiresTailwind } from '@/components/Requires'
-import HoverAnimations from '@/components/cards_temp/HoverAnimationCards'
+import HoverAnimations from '@/components/cards/HoverAnimationCards'
 import { HoverJiggleButton } from '@/showcase/animations/hover/HoverJiggleButton'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
